Fix the header for the summer log file without the set point temperature and knob position.
Also add commented out lines that can be used to plot of subset of the data.  This was added in the process of debugging.  The log file had a bunch of null characters that it probably add when the power went out and it died.  I should switch to a proper database to avoid this kind of problem.